Add `meta` argument to track arbitrary data
It can be desirable to attach additional information to a `buildImage`
output that isn't reflected in the OCI manifest. An example use-case is
constructing shadow graphs on trees of images where parts of `Env`
configurations are passed to downstream consumers. This allows
minimizing closure size while maintaining clear set/superset hierarchies
between images. For instance:

```nix
base = buildImage let
    Env = ["A=foo"];
    ExposedEnv = ["B=bar"];
in {
    Env = Env + ExposedEnv;
    meta.Env = ExposedEnv;
};

superset = buildImage {
    # Reuses derivations from `B`, but not from `A`.
    Env = ["C=baz"] + base.meta.Env;
};
```

The semantics for `meta` are the same as in `nix2container` to ease
migration to `nix-snapshotter`.